+++ src/examples/zkapps/dex/run.ts
@@ -13,7 +13,10 @@ import { expect } from 'expect';
 await isReady;
 let doProofs = false;
 
-let Local = Mina.LocalBlockchain({ proofsEnabled: doProofs });
+let Local = Mina.LocalBlockchain({
+  proofsEnabled: doProofs,
+  enforceTransactionLimits: false,
+});
 Mina.setActiveInstance(Local);
 let accountFee = Mina.accountCreationFee();
 let [{ privateKey: feePayerKey }] = Local.testAccounts;
@@ -37,12 +40,16 @@ await TokenContract.compile();
 await main({ withVesting: false });
 
 // swap out ledger so we can start fresh
-Local = Mina.LocalBlockchain({ proofsEnabled: doProofs });
+Local = Mina.LocalBlockchain({
+  proofsEnabled: doProofs,
+  enforceTransactionLimits: false,
+});
 Mina.setActiveInstance(Local);
 [{ privateKey: feePayerKey }] = Local.testAccounts;
 feePayerAddress = feePayerKey.toPublicKey();
 
 await main({ withVesting: true });
+
 console.log('all dex tests were successful! 🎉');
 
 async function main({ withVesting }: { withVesting: boolean }) {